+++ src/app/pages/Studio/components/ContextMenu/index.ts
@@ -1,1 +1,2 @@
-export { ContextMenu } from './ContextMenu';
+export * from './constants';
+export * from './ContextMenu';
