+++ src/app/pages/Studio/components/index.ts
@@ -1,6 +1,6 @@
-export { Flows } from './Flows';
-export { EndNode } from './EndNode';
-export { StartNode } from './StartNode';
-export { ContextMenu } from './ContextMenu';
-export { StandardNode } from './StandardNode';
-export { ExplorerHandle } from './ExplorerHandle';
+export * from './Flows';
+export * from './EndNode';
+export * from './StartNode';
+export * from './ContextMenu';
+export * from './StandardNode';
+export * from './ExplorerHandle';
