+++ src/app/pages/Studio/containers/Canvas/helpers.ts
@@ -6,7 +6,7 @@ import { selectFlowId } from '@/store/studio';
 import { useAppSelector, useGetNodes } from '@/hooks';
 
 import { CanvasContext } from '../../contexts';
-import { MENU_ID } from '../../components/ContextMenu/constants';
+import { MENU_ID } from '../../components';
 
 export const usePrepareHook = () => {
   const flowId = useAppSelector(selectFlowId);
